feat: implemented login processing
* fix: updated field types, fixed entity relationships and repositories

Changed data types for several entity fields, corrected JPA relationships, and fixed repository definitions to ensure proper integration with Spring Data JPA.

* feat: Seed initial RBAC data

Populated `roles`, `authorities`, `users` and `roles_authorities` tables with sample records for testing and development.

* feat: Implement basic security configuration

Added `SecurityConfig` for core security settings and `AuthUserDetailsService` for user details retrieval.

* feat: Add login template and error handling

Introduced a dedicated login page template and implemented handling for invalid credential errors during authentication.

* update: change links into layout.ftlh

* feat: added button for logout

* fix: fixed some settings into SecurityConfig

+++ src/main/java/kg/edu/mathbilim/repository/MediaTypeRepository.java
@@ -3,5 +3,5 @@ package kg.edu.mathbilim.repository;
 import kg.edu.mathbilim.model.MediaType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MediaTypeRepository extends JpaRepository<MediaType, Long> {
+public interface MediaTypeRepository extends JpaRepository<MediaType, Integer> {
 }
